Initial commit of the SmugMug Entities
This change contains the initial set of types and their properties and types that were generated based on the infomation gathered from the SmugMug API.

The current approach is to have the following types of files:
- <EntityName>.properties.cs to capture all the behavior around properties.
- <EntityName>.methods.cs to capture all the behavior that methods need to have
- <EntityName>.manual.cs for any code that needs to survive regenerating the methods and properties.

All entities derive from the SmugMugEntity class. That type is currently empty but this is where the common functionality for the object will be.