Ensure that enums that contain non-C# characters can be deserialized by JSON.NET
This requires that each enum where the value that is coming in off the wire is different than the type name be annotated with the EnumMemberAttribute.

We need to keep track of these differences during the enum analysis phase to make sure that we can correctly emit the attribute.

Also, for readability, the enum values are now generated one per line.